refactor: ui 코드 section 기준 파일 분리
